feat: write a possible way to go about generalization for reference
