+++ src/lib/components/ProjectCard.svelte
@@ -1,9 +1,7 @@
 <script lang="ts">
     export let src: string;
-    export let state: string;
     export let type: string;
     export let title: string;
-    export let dates: string;
     export let description: string;
     export let url: string;
 </script>
@@ -13,9 +11,8 @@
 >
     <img {src} alt={title} loading="lazy" class="m-0 h-64 shrink-0 object-cover md:h-48" />
     <div class="m-3 flex h-full flex-col justify-between gap-2 overflow-hidden">
-        <p class="m-0">{state} - {type}</p>
+        <p class="m-0">{type}</p>
         <h2 class="m-0">{title}</h2>
-        <p class="m-0">{dates}</p>
         <p class="m-0 grow overflow-hidden">{description}</p>
         <p class="m-0"><a href={url}>More Info</a></p>
     </div>
