+++ src/lib/projects/projects.ts
@@ -1,15 +1,21 @@
-import type { Event } from '$lib/components/EventsPanel.svelte';
 import lino from '$lib/lino-sablay.svg';
+
+export interface Project {
+    tag: string;
+    type: string;
+    title: string;
+    description: string;
+    src: string;
+    url: string;
+}
 
 export default [
     {
         tag: 'Service',
-        state: 'Past',
         type: 'Internal',
         title: 'Service Project',
-        dates: 'August 4 - August 6, 2023',
         description: 'Lorep ipsum.',
         src: lino,
         url: '/projects/',
     },
-] satisfies Event[];
+] satisfies Project[];
