+++ src/routes/projects/+page.svelte
@@ -1,10 +1,11 @@
 <script>
-    import EventsPanel from '$lib/components/EventsPanel.svelte';
-    import projects from './projects';
+    import FilterPagePanel from '$lib/components/FilterPagePanel.svelte';
+    import ProjectCard from '$lib/components/ProjectCard.svelte';
+    import projects from '$lib/projects/projects';
 
     const filters = ['All', 'Service', 'Innovation', 'Other'];
 </script>
 
 <article class="prose max-w-full dark:prose-invert">
-    <EventsPanel {filters} events={projects} />
+    <FilterPagePanel {filters} cardComponent={ProjectCard} cardsInfo={projects} />
 </article>
